pyproject.toml
[project]
name = "discord-trash-notifier"
version = "0.1.0"
description = "Discord trash notification via GitHub Actions"
readme = "discord-trash-notifier-design.md"
requires-python = ">=3.11"
dependencies = [
  "requests>=2.31.0",
  "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "time-machine>=2.14.2",
]

[tool.uv]
package = false

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
